abstract handle logger
