+++ polog/loggers/handle/abstract.py
@@ -1,0 +1,32 @@
+import functools
+from polog.core.levels import Levels
+
+
+class AbstractHandleLogger:
+    _allowed_types = {
+        'function': lambda x: type(x) is str or callable(x), # Функция, событие в которой логируется. Ожидается либо сам объект функции, либо строка с ее названием.
+        'module': lambda x: type(x) is str, # Модуль, событие в котором логируется. Ожидается только название.
+        'message': lambda x: type(x) is str, # Сообщение лога, любая строка.
+        'exception': lambda x: type(x) is str or isinstance(x, Exception), # Экземпляр перехваченного пользователем исключения или его название. Если передается экземпляр, поля с названием исключения и его сообщением будут заполнены автоматически.
+        'vars': lambda x: type(x) is str, # Ожидается любая строка, но для совместимости формата с автоматическими логами рекомендуется передавать аргументы в функцию polog.utils.json_vars(), а уже то, что она вернет, передавать сюда в качестве аргумента.
+        'success': lambda x: type(x) is bool, # Успех / провал операции, которая логируется.
+        'level': lambda x: type(x) is str or type(x) is int, # Уровень важности лога.
+        'local_variables': lambda x: type(x) is str, # json с локальными переменными, либо строка от пользователя в свободном формате.
+    }
+    _convert_values = {
+        'function': lambda x: x if isinstance(x, str) else x.__name__,
+        'exception': lambda x: x if isinstance(x, str) else type(x).__name__,
+        'level': Levels.get,
+    }
+    _convert_keys = {
+        'vars': 'local_variables',
+    }
+
+    def __getattribute__(self, name):
+        """
+        Экземпляр класса BaseLogger можно вызывать как непосредственно, так и через "методы", названия которых соответствуют зарегистрированным пользователем уровням логирования.
+        Если использован второй вариант, уровень логирования подхватится автоматически.
+        """
+        if name in Levels.levels:
+            return functools.partial(self, level=name)
+        return object.__getattribute__(self, name)
